Refactor: 서비스단 di 생성자 주입 수정 완료
