+++ src/main/java/shop/mtcoding/job/service/EnterpriseService.java
@@ -2,7 +2,6 @@ package shop.mtcoding.job.service;
 
 import java.security.NoSuchAlgorithmException;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
