+++ src/main/java/shop/mtcoding/job/service/ResumeService.java
@@ -1,20 +1,19 @@
 package shop.mtcoding.job.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import lombok.RequiredArgsConstructor;
 import shop.mtcoding.job.dto.resume.ResumeReqDto.SaveResumeReqDto;
 import shop.mtcoding.job.dto.resume.ResumeReqDto.UpdateResumeReqDto;
 import shop.mtcoding.job.handler.exception.CustomApiException;
 import shop.mtcoding.job.model.resume.Resume;
 import shop.mtcoding.job.model.resume.ResumeRepository;
 
+@RequiredArgsConstructor
 @Service
 public class ResumeService {
-
-    @Autowired
     private ResumeRepository resumeRepository;
 
     @Transactional
